- Borderlands.js, Borderlands2.js, BorderlandsPreSequel.js: Keyboard player: Added a tick that updates with 1000 Hz refresh rate to keep the user's mouse inside the game screen. - CallOfDutyBlackOps.js, CallOfDutyBlackOps2Zombies.js: Initial files - MainForm.cs: Double precision Thread.Sleep (to be able to sleep less than 1 ms) - GenericGameHandler.cs: Everything related to time is now in double-precision. Added checks for null objects that could be caused by certain games not using all available features.  * Fixed player id not being atached corrrectly  * Reverted to extracting Steam Emu to the steam emu folder (not inside game).  * Implemented Tick system: Generic Game asks us to tick the mentioned method a certain number of times during the game run.  * CenterCursor(): Clips the mouse of the keyboard player inside his game window  * Now HWnd stuff is updated only once a second - GenericGameInfo.cs: Added HandlerInterval so games can modify it, removed old variable called Interval (was integer) - IGenericGameInfo.cs: Now needs to provide a HandlerInterval double - IGameHandler.cs: Updated all time to double - PlayerInfo.cs: Added a variable holder for Player ID

+++ Master/NucleusGaming/Coop/IGameHandler.cs
@@ -15,7 +15,7 @@ namespace Nucleus.Gaming
         /// The update interval to call Update(). Set to 0 or -1
         /// to disable 
         /// </summary>
-        int TimerInterval { get; }
+        double TimerInterval { get; }
 
         ///bool Initialize(string gameFilename, List<PlayerInfo> players, Dictionary<string, GameOption> options, List<Control> addSteps, int titleHeight);
 
@@ -23,7 +23,7 @@ namespace Nucleus.Gaming
 
 
         string Play();
-        void Update(int delayMS);
+        void Update(double delayMS);
 
         void End();
